fx: move ChainBlock
