+++ meson.build
@@ -90,9 +90,10 @@ sources = [
   'src/game/control_util.c',
   'src/game/demo.c',
   'src/game/draw.c',
+  'src/game/effect_routines/bubbles.c',
+  'src/game/effect_routines/chain_block.c',
   'src/game/effects/blood.c',
   'src/game/effects/exploding_death.c',
-  'src/game/effect_routines/bubbles.c',
   'src/game/fmv.c',
   'src/game/game.c',
   'src/game/gamebuf.c',
@@ -145,7 +146,6 @@ sources = [
   'src/game/objects/effects/blood.c',
   'src/game/objects/effects/body_part.c',
   'src/game/objects/effects/bubble.c',
-  'src/game/objects/effects/chain_block.c',
   'src/game/objects/effects/dino_stomp.c',
   'src/game/objects/effects/earthquake.c',
   'src/game/objects/effects/explosion.c',
