+++ src/game/effect_routines/chain_block.c
@@ -1,10 +1,10 @@
-#include "game/objects/effects/chain_block.h"
+#include "game/effect_routines/chain_block.h"
 
 #include "config.h"
 #include "game/sound.h"
 #include "global/vars.h"
 
-void ChainBlock(ITEM_INFO *item)
+void FX_ChainBlock(ITEM_INFO *item)
 {
     if (g_Config.fix_tihocan_secret_sound) {
         Sound_Effect(SFX_LARA_SPLASH, NULL, SPM_NORMAL);
