+++ src/game/effect_routines/chain_block.h
@@ -2,4 +2,4 @@
 
 #include "global/types.h"
 
-void ChainBlock(ITEM_INFO *item);
+void FX_ChainBlock(ITEM_INFO *item);
