+++ src/global/vars.c
@@ -1,5 +1,5 @@
 #include "game/effect_routines/bubbles.h"
-#include "game/objects/effects/chain_block.h"
+#include "game/effect_routines/chain_block.h"
 #include "game/objects/effects/dino_stomp.h"
 #include "game/objects/effects/earthquake.h"
 #include "game/objects/effects/explosion.h"
@@ -19,10 +19,10 @@ char *GameMemoryPointer = NULL;
 int32_t g_FPSCounter = 0;
 
 void (*g_EffectRoutines[])(ITEM_INFO *item) = {
-    Turn180,    DinoStomp, LaraNormal,    FX_Bubbles,   FinishLevel,
-    EarthQuake, Flood,     RaisingBlock,  Stairs2Slope, DropSand,
-    PowerUp,    Explosion, LaraHandsFree, FxFlipMap,    LaraDrawRightGun,
-    ChainBlock, Flicker,
+    Turn180,       DinoStomp, LaraNormal,    FX_Bubbles,   FinishLevel,
+    EarthQuake,    Flood,     RaisingBlock,  Stairs2Slope, DropSand,
+    PowerUp,       Explosion, LaraHandsFree, FxFlipMap,    LaraDrawRightGun,
+    FX_ChainBlock, Flicker,
 };
 
 int16_t g_SampleLUT[MAX_SAMPLES] = { 0 };
